ContributionCard refactored
