+++ src/components/composite/common/cards/index.tsx
@@ -1,5 +1,5 @@
-export { default as ContributionCardV2 } from './ContributionCard/ContributionCardV2';
-export * from './ContributionCard/ContributionCardV2';
+export { default as ContributionCardV2 } from '../../../../domain/shared/components/ContributuionCard/ContributionCardV2';
+export * from '../../../../domain/shared/components/ContributuionCard/ContributionCardV2';
 
 export { default as CredentialCard } from './CredentialCard/CredentialCard';
